requirements.txt
pygame==2.6.1
pygame-ce==2.5.6
pygame_gui==0.6.14
python-i18n==0.3.9
PyYAML==6.0.3
